+++ src/components/Portal/__stories__/Portal.stories.tsx
@@ -1,4 +1,4 @@
-import type {Meta, StoryFn} from '@storybook/react';
+import type {Meta, StoryFn} from '@storybook/react-webpack5';
 
 import {Portal} from '../Portal';
 import type {PortalProps} from '../Portal';