+++ src/components/layout/Col/Col.tsx
@@ -12,7 +12,28 @@ import './Col.scss';
 
 const b = block('col');
 
-export interface ColProps extends MediaPartial<ColSize>, QAProps {
+export interface ColProps extends QAProps {
+    /**
+     * @deprecated Use "size" prop. See https://preview.gravity-ui.com/uikit/?path=/docs/components-layout--docs#col
+     */
+    s?: MediaPartial<ColSize>['s'];
+    /**
+     * @deprecated Use "size" prop. See https://preview.gravity-ui.com/uikit/?path=/docs/components-layout--docs#col
+     */
+    m?: MediaPartial<ColSize>['m'];
+    /**
+     * @deprecated Use "size" prop. See https://preview.gravity-ui.com/uikit/?path=/docs/components-layout--docs#col
+     */
+    l?: MediaPartial<ColSize>['l'];
+    /**
+     * @deprecated Use "size" prop. See https://preview.gravity-ui.com/uikit/?path=/docs/components-layout--docs#col
+     */
+    xl?: MediaPartial<ColSize>['xl'];
+    /**
+     * @deprecated Use "size" prop. See https://preview.gravity-ui.com/uikit/?path=/docs/components-layout--docs#col
+     */
+    xxl?: MediaPartial<ColSize>['xxl'];
+    size?: ColSize | [ColSize | undefined, MediaPartial<ColSize>] | MediaPartial<ColSize>;
     className?: string;
     style?: React.CSSProperties;
     children?: React.ReactNode;
@@ -22,34 +43,60 @@ export interface ColProps extends MediaPartial<ColSize>, QAProps {
  * How many columns of you 12-th column layout will take content.
  * Must be used as a child of `Row` component.
  *
- * By default component takes all available space.
- * If you wont to specify static size to all media queries use `s` prop. In mobile first layout grid is first passible value.
+ * By default, component takes all available space.
+ * If you want to specify static size use `size` prop.
  *
  * ```tsx
- * <Col s="12">some content</Col>
+ * <Col size="6">some content</Col>
+ * ```
+ * If you want responsive column use provide media sizes.
+ *
+ * ```tsx
+ * <Col size={[12, {m: 6}]}>some content</Col>
  * ```
  * ---
  *
- * Note: you can use this empty component for spacing:
+ * Note: you can use empty <Col/> component for spacing:
  *
  * ```tsx
  * <Row>
- *  <Col s="2" l="1">col 2</Col>
- *  <Col />
- *  <Col s="2" l="1">col 2</Col>
+ *   <Col size="4">col 1</Col>
+ *   <Col/>
+ *   <Col size="4">col 2</Col>
  * </Row>
  * ```
  * ---
- * Storybook - https://preview.gravity-ui.com/uikit/?path=/docs/layout--playground#col
+ * Storybook - https://preview.gravity-ui.com/uikit/?path=/docs/components-layout--docs#col
  */
-export const Col = ({children, style, className, qa, ...media}: ColProps) => {
+export const Col = ({size, children, style, className, qa, ...mediaConfigProp}: ColProps) => {
     const {getClosestMediaProps} = useLayoutContext();
-    const sizeModValue = getClosestMediaProps(media);
+
+    let mediaConfig: MediaPartial<ColSize>;
+    let defaultSizeMod: ColSize | undefined;
+
+    if (Array.isArray(size)) {
+        [defaultSizeMod, mediaConfig] = size;
+    } else if (typeof size === 'object') {
+        mediaConfig = size || mediaConfigProp;
+    } else {
+        defaultSizeMod = size;
+        mediaConfig = mediaConfigProp;
+    }
+
+    const sizeModValue = getClosestMediaProps(mediaConfig);
 
     return (
         <div
             style={style}
-            className={b({size: sizeModValue ? makeCssMod(sizeModValue) : undefined}, className)}
+            className={b(
+                {
+                    size:
+                        typeof sizeModValue === 'undefined'
+                            ? defaultSizeMod
+                            : makeCssMod(sizeModValue),
+                },
+                className,
+            )}
             data-qa={qa}
         >
             {children}