+++ src/components/layout/Container/__stories__/Container.stories.tsx
@@ -1,17 +1,16 @@
-import type {Meta, StoryFn} from '@storybook/react';
+import type {Meta, StoryObj} from '@storybook/react-webpack5';
 
 import {Col} from '../../Col/Col';
 import {Row} from '../../Row/Row';
 import {Box, LayoutPresenter} from '../../demo';
 import {Container} from '../Container';
-import type {ContainerProps} from '../Container';
 
-export default {
+const meta = {
     title: 'Components/Layout/Container',
     component: Container,
     parameters: {
         a11y: {
-            element: '#storybook-root',
+            context: '#storybook-root',
             config: {
                 rules: [
                     {
@@ -22,28 +21,30 @@ export default {
             },
         },
     },
-} as Meta;
+} satisfies Meta<typeof Container>;
+export default meta;
 
-const DefaultTemplate: StoryFn<ContainerProps> = (args) => (
-    <LayoutPresenter title="Change screen size to see different row spacing">
-        <Container {...args}>
-            <Row space="5">
-                <Col>
-                    <Box>Row</Box>
-                </Col>
-            </Row>
-            <Row space="5">
-                <Col>
-                    <Box>Row</Box>
-                </Col>
-            </Row>
-        </Container>
-    </LayoutPresenter>
-);
+type Story = StoryObj<typeof meta>;
 
-export const Default = DefaultTemplate.bind({});
-
-Default.args = {
-    spaceRow: {m: '1'},
-    maxWidth: 'l',
-};
+export const Default = {
+    render: (args) => (
+        <LayoutPresenter title="Change screen size to see different row spacing">
+            <Container {...args}>
+                <Row space="5">
+                    <Col>
+                        <Box>Row</Box>
+                    </Col>
+                </Row>
+                <Row space="5">
+                    <Col>
+                        <Box>Row</Box>
+                    </Col>
+                </Row>
+            </Container>
+        </LayoutPresenter>
+    ),
+    args: {
+        spaceRow: {s: 1, m: 2, l: 3},
+        maxWidth: 'l',
+    },
+} satisfies Story;